Add handling for ` expansion
The reader won't actually macroexpand backtick, instead it generates the
list (backtick ...) for a userland backtick implementation ala
bbloom's. While this is a little complex in that it assumes a backtick
implementation will be present when the form is evaluated, I think this
is an acceptable power to weight ratio in terms of reader simplicity.