Add a global type
The global type is the scope in which namespaces are interred so that
it's possible to do symbol resolution across namespaces and to do real
code loading. Load/Require will simply alter the current continuation by
threading a new (empty) namespace through the evaluation of the code
being loaded. The current continiation is then rebased atop the
resulting continuation and evaluation resumes. That's the plan at least.