Rework namespace system entirely
Previously, the idea was that the environment stack would look
something like this:

  Base : has special forms for bindings
  Global : rooted on Base, has Namespaces for bindings
  Namespace : rooted on Global, has Defs for bindings

The above are all conceptually global interpreter state. They
aren't thread shared at all (see not being mutable) but they
package state which would be thread shared in Clojure. Now we
get to the "local" state for the interpreter.

  Binding is either of Local or Dynamic, rooted on Namespace
  Local : Has unqualified symbols for bindings
  Dynamic : Has qualified dynamic flagged symbols for bindings.

This works... sorta. The problem comes with being logically "in"
this namespace or that while loading code. The update scope of
new defs isn't obvious, nor is the symbol resolution logic
obvious because at any given point in time you are "in" a
namespace and simultaneously updating others.

This commit simplifies and to reworks the binding tree to deal with
this complexity. Namespaces are now first class bound values,
describing all the defs and imports they contain. Namespaces and
Defs are all bound as values in the Global level namespace. This
solves the various issues of having two "places" where a Def could
be either interred or found. The price is that a "dynamic" var,
`ox.lang.bootstrap/*ns*` roughly akin to `clojure.core/*ns*` is now
requried in order to track the logically "current" namespace.
However as `*ns*` is bound with local immutable Dynamic binding
instances this is not a serious problem because no thread can leak
it's alterations thereto, nor need alterations thereto be leaked by
nested code loading.